refactor: clean up dead code from trim editor iteration
- Remove vestigial videoElement state, isEditing and onVideoElementReady
  props from asset-viewer, video-native-viewer, video-wrapper-viewer
  (superseded by TrimArea having its own video element)
- Remove unused activeHandle state from TrimManager
- Remove stale getJobCounts mocks from tests (concurrent check was removed)
- Extract duplicate formatTime into TrimManager.formatTime static method
- Restore native video controls attribute (no longer conditional)
