+++ server/src/services/asset.service.spec.ts
@@ -1593,14 +1593,6 @@ describe(AssetService.name, () => {
         audioStreams: [{}],
         format: {},
       } as any);
-      mocks.job.getJobCounts.mockResolvedValue({
-        active: 0,
-        completed: 0,
-        failed: 0,
-        delayed: 0,
-        waiting: 0,
-        paused: 0,
-      });
       mocks.assetEdit.replaceAll.mockResolvedValue([]);
 
       await sut.editAsset(authStub.admin, assetId, {
@@ -1690,14 +1682,6 @@ describe(AssetService.name, () => {
         audioStreams: [{}],
         format: {},
       } as any);
-      mocks.job.getJobCounts.mockResolvedValue({
-        active: 0,
-        completed: 0,
-        failed: 0,
-        delayed: 0,
-        waiting: 0,
-        paused: 0,
-      });
 
       await expect(
         sut.editAsset(authStub.admin, assetId, {
@@ -1828,14 +1812,6 @@ describe(AssetService.name, () => {
         audioStreams: [],
         format: {},
       } as any);
-      mocks.job.getJobCounts.mockResolvedValue({
-        active: 0,
-        completed: 0,
-        failed: 0,
-        delayed: 0,
-        waiting: 0,
-        paused: 0,
-      });
 
       await expect(
         sut.editAsset(authStub.admin, assetId, {
