+++ web/src/lib/components/asset-viewer/asset-viewer.svelte
@@ -103,8 +103,6 @@
 
   let previewStackedAsset: AssetResponseDto | undefined = $state();
   let stack: StackResponseDto | null = $state(null);
-  let videoElement = $state<HTMLVideoElement | undefined>();
-
   const asset = $derived(previewStackedAsset ?? cursor.current);
   const nextAsset = $derived(cursor.nextAsset);
   const previousAsset = $derived(cursor.previousAsset);
@@ -549,13 +547,11 @@
         cacheKey={asset.thumbhash}
         projectionType={asset.exifInfo?.projectionType}
         loopVideo={$slideshowState !== SlideshowState.PlaySlideshow}
-        isEditing={assetViewerManager.isShowEditor}
         onPreviousAsset={() => navigateAsset('previous')}
         onNextAsset={() => navigateAsset('next')}
         onClose={closeViewer}
         onVideoEnded={() => navigateAsset()}
         onVideoStarted={handleVideoStarted}
-        onVideoElementReady={(el) => (videoElement = el)}
         {playOriginalVideo}
       />
     {/if}
