+++ web/src/lib/components/asset-viewer/video-wrapper-viewer.svelte
@@ -11,13 +11,11 @@
     cacheKey: string | null;
     loopVideo: boolean;
     playOriginalVideo: boolean;
-    isEditing?: boolean;
     onClose?: () => void;
     onPreviousAsset?: () => void;
     onNextAsset?: () => void;
     onVideoEnded?: () => void;
     onVideoStarted?: () => void;
-    onVideoElementReady?: (element: HTMLVideoElement) => void;
   }
 
   let {
@@ -27,13 +25,11 @@
     cacheKey,
     loopVideo,
     playOriginalVideo,
-    isEditing = false,
     onPreviousAsset,
     onClose,
     onNextAsset,
     onVideoEnded,
     onVideoStarted,
-    onVideoElementReady,
   }: Props = $props();
 
   const effectiveAssetId = $derived(assetId ?? asset.id);
@@ -47,12 +43,10 @@
     {cacheKey}
     assetId={effectiveAssetId}
     {playOriginalVideo}
-    {isEditing}
     {onPreviousAsset}
     {onNextAsset}
     {onVideoEnded}
     {onVideoStarted}
     {onClose}
-    {onVideoElementReady}
   />
 {/if}
